src/backend/requirements.txt
fastapi
uvicorn
databases
motor
python-dotenv